+++ classes/Music/genre.rb
@@ -1,17 +1,17 @@
-require_relative '../item.rb'
+require_relative '../item'
 
 class Genre
-    attr_accessor  :name
-    attr_reader :id, :items
+  attr_accessor :name
+  attr_reader :id, :items
 
-    def initialize(name)
-        @id = Random.rand(1...1000)
-        @name = name
-        @items = []
-    end
+  def initialize(name)
+    @id = Random.rand(1...1000)
+    @name = name
+    @items = []
+  end
 
-    def add_item(item)
-        @items << item
-        item.genre = self
-    end
+  def add_item(item)
+    @items << item
+    item.genre = self
+  end
 end
